Refactored
 - renamed Triangulation.js
 - fixed nodes to array convertion in elimination func
